LPS-106022 Move AllowEditor creation to getAllowEditorProcessor
